Query database once for all users (#2261)
* Retrieve EventInstance once for all users

Moved the call to ActiveRecord EventInstance out of the loop within
KarmaCalculator to prevent hitting the database for each user.

This will relieve the workload the database needs to do in order for us
to calculate a user's karma more efficiently. There have been times
where the server kills the task due to the rake task taking to long.

* Add events parameter to KarmaCalculator

KarmaCalculator now takes two parameters now in order to decrease the
workload on the RDBMS.

* Update spec to pass two arguments to KarmaCalculator

KarmaCalculator now takes two arguments in order to reduce the worklaod
on the RDBMS.

The events parameter is expected to be an ActiveRecord query therefore
passing it a list of event_instances is necessary.

* Refactor duplicate method calls

Removed the call to more_than_one? and refactored the two calls to
convert to hash into one call.

Checking for nil events are no longer necessary since we have the RDBMS
do the heavy lifting to filter that out before KarmaCalculator is
called.

* Add exception handling to KarmaCalculations

Wrapped KarmaCalculator into a rescue clause.

This ensures the karma calculations completes for all users in the case
that there is a standard error. This will also allow us to log errors if
one occurrs while processing a particular user's karma by logging the
user's id along with the error message and backtrace.